Upgrade example project to React 0.20.0 Ported to Android
