+++ example/ProgressChart/package.json
@@ -7,7 +7,7 @@
   },
   "dependencies": {
     "art": "^0.10.0",
-    "react-native": "^0.13.2",
-    "react-native-circular-progress": "0.0.1"
+    "react-native": "^0.20.0",
+    "react-native-circular-progress": "latest"
   }
 }
