+++ src/CircularProgress.js
@@ -1,18 +1,23 @@
-import React, { View, PropTypes } from 'react-native';
+import React, { View, PropTypes, Platform } from 'react-native';
 import { Surface, Shape, Path, Group } from '../../react-native/Libraries/ART/ReactNativeART';
 import MetricsPath from 'art/metrics/path';
 
-
 export default class CircularProgress extends React.Component {
 
-  circlePath(cx, cy, r) {
+  circlePath(cx, cy, r, startDegree, endDegree) {
 
-    return Path()
-      .moveTo(cx, cx)
-      .move(r, 0)
-      .arc(r * -2, 0, r, r)
-      .arc(r * 2, 0, r, r)
-      .close();
+    let p = Path();
+    if (Platform.OS === 'ios') {
+      p.path.push(0, cx + r, cy);
+      p.path.push(4, cx, cy, r, startDegree * Math.PI / 180, endDegree * Math.PI / 180, 1);
+    } else {
+      // For Android we have to resort to drawing low-level Path primitives, as ART does not support 
+      // arbitrary circle segments. It also does not support strokeDash.
+      // Furthermore, the ART implementation seems to be buggy/different than the iOS one.
+      // MoveTo is not needed on Android 
+      p.path.push(4, cx, cy, r, startDegree * Math.PI / 180, (startDegree - endDegree) * Math.PI / 180, 0);
+    }
+    return p;
   }
 
   extractFill(fill) {
@@ -27,9 +32,10 @@ export default class CircularProgress extends React.Component {
 
   render() {
     const { size, width, tintColor, backgroundColor, style, rotation, children } = this.props;
+    const backgroundPath = this.circlePath(size / 2, size / 2, size / 2 - width / 2, 0, 360);
 
-    const circlePath = this.circlePath(size / 2, size / 2, size / 2 - width / 2);
     const fill = this.extractFill(this.props.fill);
+    const circlePath = this.circlePath(size / 2, size / 2, size / 2 - width / 2, 0, 360 * fill / 100);
 
     return (
       <View style={style}>
@@ -37,14 +43,13 @@ export default class CircularProgress extends React.Component {
           width={size}
           height={size}>
           <Group rotation={rotation - 90} originX={size/2} originY={size/2}>
+            <Shape d={backgroundPath}
+                   stroke={backgroundColor}
+                   strokeWidth={width}/>
             <Shape d={circlePath}
-              stroke={backgroundColor}
-              strokeWidth={width} />
-            <Shape d={circlePath}
-              stroke={tintColor}
-              strokeCap="butt"
-              strokeDash={[(size - width) * Math.PI * fill / 100, 9999]}
-              strokeWidth={width} />
+                   stroke={tintColor}
+                   strokeWidth={width}
+                   strokeCap="butt"/>
           </Group>
         </Surface>
         {
